+++ src/main/java/insane96mcp/survivalreimagined/mixin/AbstractArrowMixin.java
@@ -1,5 +1,6 @@
 package insane96mcp.survivalreimagined.mixin;
 
+import insane96mcp.insanelib.base.Feature;
 import insane96mcp.survivalreimagined.module.combat.feature.Stats;
 import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.phys.EntityHitResult;
@@ -12,7 +13,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(AbstractArrow.class)
 public abstract class AbstractArrowMixin {
-    @Shadow public abstract void setCritArrow(boolean p_36763_);
+    @Shadow public abstract boolean isCritArrow();
 
     @Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;hurt(Lnet/minecraft/world/damagesource/DamageSource;F)Z", shift = At.Shift.AFTER), method = "onHitEntity")
     public void onHitEntity(EntityHitResult entityHitResult, CallbackInfo ci) {
@@ -22,6 +23,8 @@ public abstract class AbstractArrowMixin {
 
     @Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/projectile/AbstractArrow;isCritArrow()Z", ordinal = 0), method = "onHitEntity")
     public boolean onCritArrowCheck(AbstractArrow arrow) {
-        return !Stats.disableCritArrowsBonusDamage;
+        if (Feature.isEnabled(Stats.class) && Stats.disableCritArrowsBonusDamage)
+            return false;
+        return this.isCritArrow();
     }
 }